third_party/WebKit/LayoutTests/bluetooth/requestDevice/blocklisted-service-in-optionalServices.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let expected = new DOMException('Origin is not allowed to access the ' +
                                  'service. Tip: Add the service UUID to ' +
                                  '\'optionalServices\' in requestDevice() ' +
                                  'options. https://goo.gl/HxfxSQ',
                                  'SecurityError');
  return setBluetoothFakeAdapter('BlocklistTestAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: [blocklist_test_service_uuid]}],
      optionalServices: ['human_interface_device']
    }))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService('human_interface_device'),
        expected, 'Blocklisted service not accessible.'),
      assert_promise_rejects_with_message(
        gattServer.getPrimaryServices('human_interface_device'),
        expected, 'Blocklisted services not accessible.')]));
}, 'Blocklisted UUID in optionalServices is removed and access not granted.');
</script>
